sync: Cleanup sync_common
